use abilities selector
