--- Packages/OsaurusCore/Views/Components/ToolSelectorView.swift
@@ -1,297 +1,0 @@
-//
-//  ToolSelectorView.swift
-//  osaurus
-//
-//  Tool selector showing available tools and their current state.
-//  Toggles update persona overrides (or global config for Default persona).
-//
-
-import SwiftUI
-
-struct ToolSelectorView: View {
-    /// The persona ID to update when toggling tools
-    let personaId: UUID
-
-    /// Observe ToolRegistry for live updates when tools change
-    @ObservedObject private var toolRegistry = ToolRegistry.shared
-    /// Observe PersonaManager for live updates when persona overrides change
-    @ObservedObject private var personaManager = PersonaManager.shared
-
-    @State private var searchText: String = ""
-    @Environment(\.theme) private var theme
-
-    private var tools: [ToolRegistry.ToolEntry] {
-        toolRegistry.listTools(withOverrides: personaOverrides)
-    }
-
-    private var personaOverrides: [String: Bool]? {
-        personaManager.effectiveToolOverrides(for: personaId)
-    }
-
-    private var filteredTools: [ToolRegistry.ToolEntry] {
-        if searchText.isEmpty {
-            return tools
-        }
-        return tools.filter {
-            SearchService.matches(query: searchText, in: $0.name)
-                || SearchService.matches(query: searchText, in: $0.description)
-        }
-    }
-
-    /// Count of enabled tools
-    private var enabledCount: Int {
-        tools.filter { $0.enabled }.count
-    }
-
-    /// Toggle a tool's enabled state for this persona
-    private func toggleTool(_ name: String, currentlyEnabled: Bool) {
-        personaManager.setToolEnabled(!currentlyEnabled, tool: name, for: personaId)
-    }
-
-    /// Enable all tools for this persona
-    private func enableAll() {
-        personaManager.enableAllTools(for: personaId, tools: tools.map { $0.name })
-    }
-
-    /// Disable all tools for this persona
-    private func disableAll() {
-        personaManager.disableAllTools(for: personaId, tools: tools.map { $0.name })
-    }
-
-    var body: some View {
-        VStack(spacing: 0) {
-            // Header with count
-            header
-
-            Divider()
-                .background(theme.primaryBorder.opacity(0.3))
-
-            // Search field
-            searchField
-
-            Divider()
-                .background(theme.primaryBorder.opacity(0.3))
-
-            // Tool list
-            if filteredTools.isEmpty {
-                emptyState
-            } else {
-                toolList
-            }
-        }
-        .frame(width: 360, height: min(CGFloat(tools.count * 56 + 140), 480))
-        .background(
-            RoundedRectangle(cornerRadius: 12, style: .continuous)
-                .fill(theme.primaryBackground)
-                .shadow(color: .black.opacity(0.2), radius: 16, x: 0, y: 8)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-        .overlay(
-            RoundedRectangle(cornerRadius: 12, style: .continuous)
-                .strokeBorder(theme.primaryBorder.opacity(0.3), lineWidth: 0.5)
-        )
-    }
-
-    // MARK: - Header
-
-    private var header: some View {
-        VStack(spacing: 10) {
-            // Title row
-            HStack {
-                Text("Available Tools")
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundColor(theme.primaryText)
-
-                Spacer()
-
-                Text("\(enabledCount)/\(tools.count)")
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundColor(theme.secondaryText)
-                    .padding(.horizontal, 8)
-                    .padding(.vertical, 3)
-                    .background(
-                        Capsule()
-                            .fill(theme.secondaryBackground)
-                    )
-            }
-
-            // Action buttons row
-            HStack(spacing: 8) {
-                Button(action: enableAll) {
-                    Text("Enable All")
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundColor(theme.primaryText)
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 5)
-                        .background(
-                            RoundedRectangle(cornerRadius: 6, style: .continuous)
-                                .fill(theme.secondaryBackground)
-                        )
-                }
-                .buttonStyle(.plain)
-
-                Button(action: disableAll) {
-                    Text("Disable All")
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundColor(theme.primaryText)
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 5)
-                        .background(
-                            RoundedRectangle(cornerRadius: 6, style: .continuous)
-                                .fill(theme.secondaryBackground)
-                        )
-                }
-                .buttonStyle(.plain)
-
-                Spacer()
-            }
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 12)
-    }
-
-    // MARK: - Search Field
-
-    private var searchField: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "magnifyingglass")
-                .font(.system(size: 13))
-                .foregroundColor(theme.tertiaryText)
-
-            TextField("Search tools...", text: $searchText)
-                .textFieldStyle(.plain)
-                .font(.system(size: 13))
-                .foregroundColor(theme.primaryText)
-
-            if !searchText.isEmpty {
-                Button(action: { searchText = "" }) {
-                    Image(systemName: "xmark.circle.fill")
-                        .font(.system(size: 12))
-                        .foregroundColor(theme.tertiaryText)
-                }
-                .buttonStyle(.plain)
-            }
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 8)
-        .background(theme.secondaryBackground.opacity(0.5))
-    }
-
-    // MARK: - Empty State
-
-    private var emptyState: some View {
-        VStack(spacing: 8) {
-            Image(systemName: "magnifyingglass")
-                .font(.system(size: 24))
-                .foregroundColor(theme.tertiaryText)
-            Text("No tools found")
-                .font(.system(size: 13))
-                .foregroundColor(theme.secondaryText)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding()
-    }
-
-    // MARK: - Tool List
-
-    private var toolList: some View {
-        ScrollView {
-            LazyVStack(spacing: 2) {
-                ForEach(filteredTools) { tool in
-                    ToolRowItem(
-                        tool: tool,
-                        onToggle: { toggleTool(tool.name, currentlyEnabled: tool.enabled) }
-                    )
-                }
-            }
-            .padding(.vertical, 6)
-            .padding(.horizontal, 4)
-        }
-    }
-}
-
-// MARK: - Tool Row Item
-
-private struct ToolRowItem: View {
-    let tool: ToolRegistry.ToolEntry
-    let onToggle: () -> Void
-
-    @State private var isHovered: Bool = false
-    @Environment(\.theme) private var theme
-
-    var body: some View {
-        HStack(spacing: 12) {
-            // Toggle with theme accent color
-            Toggle(
-                "",
-                isOn: Binding(
-                    get: { tool.enabled },
-                    set: { _ in onToggle() }
-                )
-            )
-            .toggleStyle(SwitchToggleStyle(tint: theme.accentColor))
-            .scaleEffect(0.7)
-            .frame(width: 36)
-
-            // Tool info
-            VStack(alignment: .leading, spacing: 3) {
-                Text(tool.name)
-                    .font(.system(size: 12, weight: .medium))
-                    .foregroundColor(tool.enabled ? theme.primaryText : theme.secondaryText)
-                    .lineLimit(1)
-
-                Text(tool.description)
-                    .font(.system(size: 10))
-                    .foregroundColor(theme.tertiaryText)
-                    .lineLimit(2)
-            }
-
-            Spacer()
-
-            // Token estimate (catalog entry size for two-phase loading)
-            HStack(spacing: 2) {
-                Text("~\(tool.catalogEntryTokens)")
-                    .font(.system(size: 10, weight: .medium, design: .monospaced))
-                    .foregroundColor(theme.tertiaryText)
-
-                Text("tokens")
-                    .font(.system(size: 9, weight: .regular))
-                    .foregroundColor(theme.tertiaryText.opacity(0.6))
-            }
-            .help(
-                "Catalog entry: ~\(tool.catalogEntryTokens) tokens. Full schema if selected: ~\(tool.estimatedTokens) tokens"
-            )
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 8)
-        .background(
-            RoundedRectangle(cornerRadius: 8, style: .continuous)
-                .fill(isHovered ? theme.secondaryBackground.opacity(0.6) : Color.clear)
-        )
-        .contentShape(Rectangle())
-        .onHover { hovering in
-            withAnimation(.easeInOut(duration: 0.1)) {
-                isHovered = hovering
-            }
-        }
-    }
-}
-
-// MARK: - Preview
-
-#if DEBUG
-    struct ToolSelectorView_Previews: PreviewProvider {
-        struct PreviewWrapper: View {
-            var body: some View {
-                ToolSelectorView(personaId: Persona.defaultId)
-                    .padding()
-                    .frame(width: 450, height: 550)
-                    .background(Color.gray.opacity(0.2))
-            }
-        }
-
-        static var previews: some View {
-            PreviewWrapper()
-        }
-    }
-#endif
